requirements.txt
pandas>=2.2,<3
numpy>=1.26,<3
pyarrow>=15,<18
PyYAML>=6.0,<7
scikit-learn>=1.4,<2
